* Remove method promotion at toplevel (causes unnecessary imports of the whole Splitgraph anytime anything is imported, since the "splitgraph" package has to get imported first) * Rename ._common packages to .common (in commandline and core) since everything seems to use them * Push most imports in splitgraph.commandline to be local so that running e.g. sgr or a single sgr command only imports required packages, not all of splitgraph. This brings the time to run `sgr` to ~200ms down from ~600ms in an empty venv (1s with pkg_resources installed)
