+++ splitgraph/config/config.py
@@ -5,7 +5,7 @@ from .argument_config import get_argument_config_value
 from .config_file_config import get_config_dict_from_config_file
 from .default_config import get_default_config_value
 from .environment_config import get_environment_config_value
-from .keys import KEYS, ALL_KEYS, ConfigDict, DEFAULTS
+from .keys import KEYS, ALL_KEYS, ConfigDict
 from .system_config import get_system_config_value
