+++ splitgraph/core/__init__.py
@@ -10,13 +10,3 @@ following methods:
   * By using :func:`splitgraph.core.engine.lookup_repository` which will search for the repository on the current
     lookup path.
 """
-
-from .engine import *
-from .registry import (
-    publish_tag,
-    get_published_info,
-    unpublish_repository,
-    get_info_key,
-    set_info_key,
-)
-from .repository import Repository, import_table_from_remote, clone
