+++ splitgraph/core/metadata_manager.py
@@ -8,8 +8,8 @@ from psycopg2.extras import Json
 from psycopg2.sql import SQL, Identifier
 
 from splitgraph.config import SPLITGRAPH_API_SCHEMA, SPLITGRAPH_META_SCHEMA
-from splitgraph.core._common import TableSchema
-from ._common import select, ResultShape
+from splitgraph.core.common import TableSchema
+from .common import select, ResultShape
 
 if TYPE_CHECKING:
     from splitgraph.core.repository import Repository
