+++ test/splitgraph/commands/test_bloom_indexing.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 import pytest
 
-from splitgraph import ResultShape
+from splitgraph.core.common import ResultShape
 from splitgraph.core.indexing.bloom import _prepare_bloom_quals, filter_bloom_index, describe
 from test.splitgraph.conftest import OUTPUT
