+++ test/splitgraph/test_security.py
@@ -2,8 +2,8 @@ import pytest
 from psycopg2._psycopg import ProgrammingError
 from psycopg2.sql import SQL, Identifier
 
-from splitgraph import SPLITGRAPH_META_SCHEMA
-from splitgraph.core._common import META_TABLES, select
+from splitgraph.config import SPLITGRAPH_META_SCHEMA
+from splitgraph.core.common import META_TABLES, select
 from splitgraph.core.registry import get_published_info, unpublish_repository
 from splitgraph.core.repository import Repository, clone
